feat(line-profile-page): translate content & beautify page
