+++ src/locale/en.json
@@ -108,7 +108,7 @@
   "funding_paragraph": "The Databus project is developed by the Hasadna for public knowledge, by the contributions of volunteers, and based upon",
   "all_rides_completed": "All/Almost the rides are completed",
   "missing_rides": "Of the rides are missing",
-  "line": "line",
+  "line": "Line",
   "from": "from",
   "destination": "destination",
   "velocity": "velocity",
@@ -118,8 +118,26 @@
   "drive_direction": "drive direction",
   "bearing": "bearing",
   "coords": "coords",
-  "hide_document": "hide document",
-  "show_document": "show document",
+  "hide_document": "Hide geeks data",
+  "show_document": "Show geeks data",
+  "lineProfile": {
+    "title": "Profile for Line",
+    "notFound": "We couldn't find the line you were looking for :(",
+    "id": "ID",
+    "date": "Date",
+    "lineReference": "Line Reference",
+    "operatorReference": "Operator Reference",
+    "agencyName": "Agency Name",
+    "route": {
+      "shortName": "Route Short Name",
+      "start": "Start",
+      "end": "End",
+      "mkt": "Route MKT",
+      "direction": "Route Direction",
+      "alternative": "Route Alternative",
+      "type": "Route Type"
+    }
+  },
   "aboutPage": {
     "title": "About us",
     "privacyText": "<0>Google Analytics</0> service is embedded in the website to analyze the usage patterns and positioning of the website in search engines. This code reveals to the operators of the service information regarding the behavior of the users. <1>Learn more</1> about the privacy policy of the service.",
