+++ src/pages/lineProfile/LineProfileDetails.scss
@@ -1,0 +1,8 @@
+table {
+  display: flex;
+  justify-content: center;
+
+  td {
+    padding: 0 1rem;
+  }
+}
