+++ src/pages/lineProfile/Route.interface.ts
@@ -1,0 +1,13 @@
+export interface Route {
+  id: number
+  date: string
+  line_ref: number
+  operator_ref: number
+  route_short_name: string
+  route_long_name: string
+  route_mkt: string
+  route_direction: string
+  route_alternative: string
+  agency_name: string
+  route_type: string
+}
